:space_invader: more progress in GA refactor
