+++ components/ExhibitionsComponents/Exhibition/Details/index.js
@@ -15,12 +15,7 @@ class Details extends React.Component {
   // Google Analytics tracking for exhibit home view event
   componentDidMount() {
     this.trackEvent();
-    Router.router.events.on("routeChangeComplete", this.trackEvent);
-  }
-
-  // Cleanup, prevents multiple pageviews being counted for a single route.
-  componentWillUnmount() {
-    Router.router.events.off("routeChangeComplete", this.trackEvent);
+    Router.onRouteChangeComplete = url => this.trackEvent();
   }
 
   trackEvent() {
