+++ components/PrimarySourceSetsComponents/Source/components/ContentAndMetadata/index.js
@@ -1,5 +1,6 @@
 import React from "react";
 import ReactMarkdown from "react-markdown";
+import Router from "next/router";
 
 import {
   ZoomableImageViewer,
@@ -7,11 +8,10 @@ import {
   VideoPlayer,
   PDFViewer
 } from "components/shared/mediaViewers";
-import Button from "components/shared/Button";
 import CiteButton from "components/shared/CiteButton";
-import GaPssWrapper from "./GaPssWrapper";
 
 import {
+  getFullPath,
   joinIfArray,
   getItemId,
   getPartner,
@@ -20,6 +20,7 @@ import {
   trackGaEvent
 } from "lib";
 import { markdownLinks } from "lib/externalLinks";
+import * as gTag from "lib/gTag";
 
 import utils from "stylesheets/utils.scss";
 import css from "./ContentAndMetadata.scss";
@@ -88,6 +89,8 @@ class ContentAndMetadata extends React.Component {
   componentDidMount() {
     // now collapse it
     this.setState({ isOpen: false });
+    this.trackSourceView();
+    Router.onRouteChangeComplete = url => this.trackSourceView();
   }
 
   componentWillReceiveProps() {
@@ -96,6 +99,24 @@ class ContentAndMetadata extends React.Component {
 
   showMoreDescription() {
     this.setState({ isOpen: true });
+  }
+
+  trackSourceView() {
+    const fullPath = getFullPath();
+    const source = this.props.source;
+
+    if (fullPath !== this.lastTrackedPath) {
+      const gaEvent = {
+        type: "View Primary Source",
+        itemId: getItemId(source),
+        title: joinIfArray(getTitle(source)),
+        partner: joinIfArray(getPartner(source)),
+        contributor: joinIfArray(getContributor(source))
+      };
+
+      gTag.event(gaEvent);
+      this.lastTrackedPath = fullPath;
+    }
   }
 
   render() {
@@ -272,4 +293,4 @@ class ContentAndMetadata extends React.Component {
   }
 }
 
-export default GaPssWrapper(ContentAndMetadata);
+export default ContentAndMetadata;
