+++ pages/exhibitions/exhibition/index.js
@@ -1,6 +1,6 @@
-import { withRouter } from "next/router";
 import React from "react";
 import fetch from "isomorphic-fetch";
+import { withRouter } from "next/router";
 
 import MainLayout from "components/MainLayout";
 import BreadcrumbsModule from "components/PrimarySourceSetsComponents/BreadcrumbsModule";
